requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain
langchain-community
chromadb
sentence-transformers
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
python-jose[cryptography]==3.3.0
pydantic==2.5.0
requests==2.31.0
pydantic-settings>=2.0.0